+++ module-domain/src/main/java/com/whiskey/domain/payment/dto/PaymentConfirmCommand.java
@@ -1,10 +1,8 @@
 package com.whiskey.domain.payment.dto;
 
-import java.math.BigDecimal;
-
 public record PaymentConfirmCommand(
     Long memberId,
     String orderId,
-    BigDecimal amount,
+    Long amount,
     String paymentKey
 ) {}
